+++ CHANGELOG.md
@@ -1,7 +1,22 @@
 # CHANGELOG - WORDLE
 
-## [1.0.0] 
+## [1.0.0] - First version of wordle
+- Changelog created.
 - Initial version of Wordle.
 - Unwinnable bug fixed.
 - Valid wordle list added.
 - Possible answers list added.
+
+## [1.1.0] - Wordle Custom 
+### New Features:
+- Thematic mode added - it allows you to play with words of specific categories.
+- Custom attempts added - the user can now set the number of guesses they want.
+### Visual changes:
+- Board and options added.
+- Improved display of HOW TO PLAY and win/lose messages.
+### Internal Changes:
+- Animals, Countries and fr-vg lists added for thematic mode.
+- Valid wordle list modified and slightly extended.
+- Fixed minor bugs.
+
+